requirements.txt
Flask==2.2.3
requests==2.28.2
pandas==1.5.3
openpyxl==3.1.2
python-dotenv==1.0.0
Werkzeug==2.2.3
gunicorn==20.1.0
Flask-WTF==1.1.1
openai>=1.0.0
googlemaps==4.10.0
beautifulsoup4==4.12.2
